UANCABOT/requirements.txt
matplotlib
pygame
pyserial
plotly
keyboard
